+++ srcs/projection.c
@@ -6,7 +6,7 @@
 /*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/15 15:41:08 by sakitaha          #+#    #+#             */
-/*   Updated: 2024/02/21 16:28:05 by sakitaha         ###   ########.fr       */
+/*   Updated: 2024/03/14 13:33:35 by sakitaha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,13 +25,13 @@ static void	orthographic_projection(t_point *point)
 	point->y_2d = point->y;
 }
 
-static void	oblique_projection(t_point *point, float degrees)
+static void	oblique_projection(t_point *point, double degrees)
 {
 	point->x_2d = point->x + point->z * cos(deg_to_rad(degrees));
 	point->y_2d = point->y + point->z * sin(deg_to_rad(degrees));
 }
 
-static void	cavalier_projection(t_point *point, float depth_scale)
+static void	cavalier_projection(t_point *point, double depth_scale)
 {
 	point->x_2d = point->x + (point->z * depth_scale) * cos(deg_to_rad(45));
 	point->y_2d = point->y + (point->z * depth_scale) * sin(deg_to_rad(45));
